Fix/ipfs rewards#60 (#61)
* update new rewards class

* create Reward instead of RewardIPFS in the SDK

* import fixed

* fixes loading from block
